Updated openstack/openstack Project: openstack-infra/system-config  6969569a380441865a31f7c2c2dd8531db967c8a
Fix logstash-filters test

In order to run the script directly in the test, the script needs to
have the executable bit set. Currently, the test is failing with a
"Permission denied" error when attempting to run the script.

We also need to run install_modules.sh in the logstash-filters test, so
move that invocation to the shared prep-apply.sh script.

The vcsrepo puppet define doesn't understand revisions of the form
"refs/zuul/foo", so instead parse the HEAD sha and pass that to vcsrepo.

Finally, we need to ensure that $elasticsearch_nodes is populated with
an array, otherwise the function @elasticsearch_nodes.map in the
output.conf template will fail.

Change-Id: Ic1ab6d04232abb3cf347f0d27afbf9172f62b0e4