Updated openstack/openstack Project: openstack/cinder  ece0ee31c6f2963e0efcbb988e3a505f396a7d41
Fix retype with migrate

When a retype requires migrating a volume the check on the volume
manager will always consider that there's no need to migrate so it will
not try to migrate the volume but retype it instead, so in drivers that
implement `retype` (like LVM) it will falsely report success.

This patch fixes the check to detect if the volume is in our own
backend.

Closes-bug: #1652802
Change-Id: I5e977ed412cf4a1614d7728ffe7898435291fcc3