Updated openstack/openstack Project: openstack/neutron-specs  9f8d91852c1760c1bec97b383abb839def935d92
Spec for providing Nova portbinding information for live migration

The goal is to provide nova the required portbinding information
for live migration. This spec proposes to allow a port to have
a second inactive binding. During migration, the host that
runs the instance has the active binding, while the binding
belonging to the other host is inactive.

A new API is required to externalize this new inactive binding.

Partial-Bug: #1580880
Co-authored-by: Brian Stajkowski <brian.stajkowski@rackspace.com>
Change-Id: I12d1310b8df30f5fbae4cb9c8e804d34c20ee7c9