Updated openstack/openstack Project: openstack/trove  88de697b814999eef7a2cb98ba8de740539f0455
Redis backup continues if auto one already running

If you kick off a backup and the internal backup is already
running, then the manual one will fail.  Since this is what
the backup command would run anyways, we should use it
and just drive on.

The bgsave command was modified to detect the running
command and just prints out a message before
continuing.

Change-Id: I43ededa0e0f60c2abc5b7a5d4981e8a97094b62c
Closes-Bug: 1656432