Updated openstack/openstack Project: openstack/django_openstack_auth  f3c21575d2a3effe1a82a8f22110ea542cb9c9dc
Fix exception catch-all in domain scope auth

Previously, the get_domain_scoped_auth plugin method caught any
exceptions found while trying to scope a token and logged the error
without addressing it. This was hiding an error that was occuring in
the unit tests, which was that some of the plugin calls were not being
mocked properly. This patch narrows down the exception handling to the
same exceptions handled in the project scoping case and adds the
necessary mocks to the tests.

Change-Id: I80a085ca731391b3f54a5ef999c92ab8ba3e69a0