Updated openstack/openstack Project: openstack/cinder  5e04f054460887a9185c6d2bc203dd913f120137
Prevent claiming and updating races on worker

Current code for claiming and updating workers relays on the updated_at
field to determine when a DB record has changed.  This is usually enough
for any DB with sub-second resolution since the likeliness of us having
a race condition is very unlikely.

But not all DBs support sub-second resolution, and in those cases the
likeliness of a race condition increases considerable since we are
working with a 1 second granularity.

This patch completely removes the possibility of having race conditions
using a specific integer field that will be increased on each DB update.
It is compatible with both types of DBMs and will also work with rolling
upgrades.

The reason why we are not using the version counting provided by
SQLAlchemy [1] is because we require an ORM instance to use the feature,
and in some of our usages we don't have an instance to work with, and
adding an additional read query is unnecessarily inefficient.

Additionally we will no longer see spurious errors in unit test
test_do_cleanup_not_cleaning_already_claimed_by_us.

[1] http://docs.sqlalchemy.org/en/latest/orm/versioning.html

Implements: blueprint cinder-volume-active-active-support
Change-Id: Ief9333a2389d98f5d0a11d8da94d160de8ecce0e