Updated openstack/openstack Project: openstack-infra/tripleo-ci  34613ce5fe505c53a12e180a827edd9e1aa41838
tripleo.sh: skip plan delete for mitaka

At the end of the overcloud delete, we delete the plan too.
The overcloud plan delete command was introduced only in the newton
cycle, so mitaka tests fail with an error.
This patch skip plan delete if we are testing overcloud delete in
mitaka.

Change-Id: Iff144f9504b15998e078bb69a01eb4d6f2cb1498
Closes-Bug: 1657762