Updated openstack/openstack Project: openstack-infra/devstack-gate  f1025531af1726e4fff54b31e16d472bae8193de
Run discover_hosts to register subnodes in a multinode setup

CellsV2 requires that compute hosts are registered with a cell.

In a single-node devstack run the 'nova-manage cell_v2 simple_cell_setup'
command will create a cell from the main (only) cell database, discover
the compute node and map it to the cell.

In a multinode setup, we need to run discover_hosts on the primary node
after the subnode is setup since the subnode doesn't know which cell
it's going to be in and doesn't (or can't rely on) have the database
connection to auto-register.

The discover_hosts command is idempotent so running this from a
single-node or multi-node setup will make no difference.

Depends-On: Icc595d60de373471aa7ee8fb9f3a81fc12d80438
Change-Id: I8938fd37f71498e87aa348a81c357eef5c1a8093