Updated openstack/openstack Project: openstack/python-zaqarclient  23aa5c1a0549ad76224de65348909cb6662e64c4
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Ieba864a99df3b16b8d3784dabe360be2e0145f0a