Updated openstack/openstack Project: openstack/tripleo-image-elements  c8543f42705b4daa5c41f4bdfdb544442793fb2b
Add a script to zero /etc/sysconfig/iptables at build time

When including this element we empty the stock /etc/sysconfig/iptables
file as shipped by the iptables rpm package. The reason for this is that
puppet firewall has a hard time to cope with exiting rules when
/etc/sysconfig/iptables is populated and the iptables service is not
active. The referenced bug has a full explanation for the problem.

Partial-Bug: #1657108

Change-Id: Iddc21316a1a3d42a1a43cbb4b9c178adba8f8db3