Updated openstack/openstack Project: openstack/neutron  2614bc649844b3a138a5b9e48c54fedd193433dc
DVR: Fix IPtables driver for metering with DVR routers

IPTables driver for metering was not handling the DVR router
namespaces properly for configuring the metering forward rules.

This patch addresses the issue by configuring the iptables
manager based on the availability of the namespace and selecting
the respective namespaces such as router namespace and snat
namespace with the right external device.

Change-Id: I6790d6ff42d9f8fa220e1a231ae94cbf8b60506c
Closes-Bug: #1506567