Updated openstack/openstack Project: openstack-infra/tripleo-ci  2ba538ecfb28516df665036322ad2877a537fa65
TLS everywhere job introduction

This introduces a job that deploys TLS everywhere (or at least what's
been developed so far).

It's an OVB job with a suffix of fakeha-caserver.

fakeha indicates that it deploys pacemaker (with network isolation),
but I labelled it as fake since it's only one controller, and not 3.

caserver indicates that this job also deploys a CA server, which in
this case is FreeIPA.

FreeIPA is deployed in an extra OVB node, using OVB's undercloud
environment; which means that this node is available when we get an
environment.

HAProxy is using TLS on all endpoints (both public and internal), and
so are most of the services (which are in the internal network).
The certificates they use are provided via certmonoger, which
requests them from FreeIPA. To do this, the nodes need to enroll,
and to do this, I have prepared a script that runs in the NodeTLSCAData
of the nodes that does the enrollment. This method of enrollment will
be replaced by a more automated approach (either via mistral or via
the novajoin vendordata plugin for nova).

Right now, this gate is running on the experimental queue, however,
the plan is to introduce it as a periodic job.

bp tls-via-certmonger

Depends-On: I44a8135118b047f154e3edbb8846751bd2e16c84
Depends-On: I22ebe028c7b23cb5fba7d03ce89e967bca5951f8
Depends-On: Ie6234a15af7aa493257c13bd40cbf51ab36085be
Depends-On: If3e6cb1dde235b24a0c188eb758f9a2e6a626f69
Depends-On: I38b880250c3b30d7fa1c9e56e82a1be8b59b7e30
Change-Id: Id5dfb38852cf2420f4195a3c1cb98d5c47bbd45e