Updated openstack/openstack Project: openstack/nova  d4f21f24dc743a6759abdd8570589e6cdd9c8a6e
Strict pattern match query parameters

We have a lot of filters are pattern match.

There is list for exact match
https://github.com/openstack/nova/blob/df2fd4a252cecc1e1ef471c071e57526ddf65499/nova/db/sqlalchemy/api.py#L2221

Out of that list will be pattern match
https://github.com/openstack/nova/blob/df2fd4a252cecc1e1ef471c071e57526ddf65499/nova/db/sqlalchemy/api.py#L2231

HTTP 500 raises if invalid regex provided for
those filters, strict there format to be regex
in JSON schema to avoid this.

partial implement of bp add-whitelist-for-server-list-filter-sort-parameters
Closes-Bug: #1653899

Change-Id: I4cf38407c20284dee7127edfe312da81caac9272