Updated openstack/openstack Project: openstack/neutron  0c05d49949b25204482180efd26be3e20310b043
Server-side push notifications for ML2

ML2 server-side push notifications for subnets, networks,
ports, security groups, and security group rules.

This adds a registry callback listener for each of the objects
above for the AFTER_CREATE and AFTER_UPDATE events. Whenever
one of these is triggered, it will parse the ID out of the event
and use it to retrieve an OVO object from the DB and dispatch
it to the RPC callback notifier.

This is only the logic to push changes to the agents. It does not
include an interface for the agents to query the server or the agent
side code to receive the notifications.

Partially-Implements: blueprint push-notifications
Change-Id: I5efc625c5e8565693e795d70e0f85810552d38cd