Updated openstack/openstack Project: openstack-infra/project-config  05a171cd30634624c51dc64ed1598886a5815118
Skip non-tripleo-quickstart OVB jobs for files they do not cover

While we iterate on tripleo-quickstart based CI, it would be nice
to not waste a ton of OVB resources on the tripleo cloud for each
patch. This updates the zuul layout to skip those jobs if we are
only touching quickstart specific files.

Change-Id: Ic5505b87e70f34ba1aafe11ecfdcce3d87e2f6f1