Updated openstack/openstack Project: openstack/ironic  078e813f5009b4fc6a87dd92fb6fc12caf6b04f1
DevStack: Configure nodes/environment to boot in UEFI mode

This patch is adding support for DevStack to configure nodes and
environment (Ubuntu and Fedora) to deploy nodes in UEFI mode.

A new configuration called IRONIC_BOOT_MODE was added to DevStack and
supports two values: "bios" and "uefi". Defaults to "bios".

Partial-Bug: #1625616
Change-Id: I4e9247c6a91c707db08247a8ac04b1f146cd3714