Updated openstack/openstack Project: openstack/glance  18acc704a1573346f751a2b8260720fd09e664cd
Fix regression introduced by Community Images

Updating an image in v1 skipped the work to ensure that the image dictionary
would be Image v2 compliant. It was hidden inside an else clause and was only
run when there was no image id provided. This meant that only sometimes would
is_public be appropriately converted to visibility.

A functional test has been added to prevent regression and the code has been
mildly altered to fix the issue.

Closes-bug: #1658164
Change-Id: I996fbed2e31df8559c025cca31e5e12c4fb76548