Updated openstack/openstack Project: openstack/cinder  7f0475f15265f8499717bfec2379a9af6f1314cd
Make notify_service_capabilities cluster aware

We recently introduced a mechanism to notify ceilometer of updated
service capabilities, but the mechanism was meant to work with the old
implementation of Active/Active where jobs were we overused the host
field for the cluster.

This patch changes the way that mechanism work and makes it aware of the
new cluster mechanism.

To do this we change what data we store in service_states dictionary in
the host_manager.  Prior to this patch we were storing all the
individual hosts, but now we store only the data per backend, so we are
either storing the capabilities with the host or the cluster as the key.

Implements: blueprint cinder-volume-active-active-support
Change-Id: I8732ce321f4d0c41b914afa7acf48d6cf839d9be