Updated openstack/openstack Project: openstack/nova  32e2654152a1b805375d8c7b5a5c41ad2ee8ce65
Remove invalid URL in gabbi tests

The use of a literal snowman in a URL is an invalid URL. In Python 3
the url parsing code used by gabbi will not accept this URL and
correctly rejects it. Python 2 is less strict and passes it on
(presumably encoding it somewhere in the stack).

Python 3's behavior is correct and the fact that the test ever
worked was an artifact of Python 2's bad behavior. The test
immediately following the removed test is testing the correct and
legitimate HTTP behavior, so it is best to just remove the bad one.

By doing so all functional tests pass under python3 so the
test-functional-py3.txt has been made empty, with a comment. This
was done, instead of removing it, to signal the state of affairs.

Change-Id: If40f5093c65491dcfc225dffd95c6dc1318a0f99