Updated openstack/openstack Project: openstack/nova  e1a4d0618df1836d3fcd04ebe3597275b7deb361
rt: use a single ResourceTracker object instance

This patch removes the ResourceTracker.nodename attribute, switches the
compute manager to only create a single ResourceTracker object instance
which itself now has an in-memory dict of ComputeNode objects that are
managed by the nova-compute daemon.

This isolates the code that manages ComputeNode objects and resources in
just the resource tracker, which will make it possible for the scheduler
report client to manage Ironic nodes and custom resource classes
properly. The scheduler report client contains a cache of
ResourceProvider, Inventory, and Allocation records. We definitely did
not want to have multiple ResourceTracker object instances, each with
their own cache of ResourceProvider objects. Having a single
ResourceTracker and single scheduler report client is both more
efficient and a cleaner interface into the compute manager.

Co-Authored-By: Chris Dent <cdent@anticdent.org>
Change-Id: I6827137f35c0cb4f9fc4c6f753d9a035326ed01b
blueprint: custom-resource-classes