Updated openstack/openstack Project: openstack/nova  78edb82d72f0eebdfaed5312ccc9c42a15bd9414
Add a PlacementFixture

Some of the functional tests are calling the scheduler for getting the destination.
Once we will accept having the scheduler be calling the Placement service in order
to know a list of valid resource providers, we will need to run a local fixture
for having those tests faking the Placement service.

Change-Id: I6883888bf5d1920ab57dbf3dcaa1a7b375591754