Updated openstack/openstack Project: openstack/swift  a1c88b906b9b056aa3f8b56be1eb000935fc0be4
Extract test pattern to helper

An existing test in diskfile established an prior art for a pattern to
create a diskfile with a different suffix - I'd like to make use of it
in new tests in multiple unrelated change sets.

Also add a test to demonstrate some existing robustness and prevent
regression.

Author: Pavel Kvasnička <pavel.kvasnicka@firma.seznam.cz>
Co-Author: Alistair Coles <alistair.coles@hpe.com>
Related-Change-Id: I64cadb1a3feb4d819d545137eecfc295389794f0
Change-Id: I3a661fae5c7cfeb2dbcdb7f46941f55244d0b9ad