Updated openstack/openstack Project: openstack/neutron  74a227500aaf9593f4b2b0e44fbf9cffda650405
Register sqlalchemy events through hook for UT cleanup

Register all sqlalchemy events through a new function in
neutron.db.api so we can keep track of active events and
ensure all are removed at the end of each test run.

Without this, an instance of a plugin may be left around
with the only reference to it existing in SQLAlchemy, where
it will receive events for tests unrelated to it and potentially
interfere.

Change-Id: I8e93eb4e8ef5a13f015db9cd20e44941cdcb72ef