Updated openstack/openstack Project: openstack-infra/shade  a616b64402c658c09632bb8b5961878cd06b0d26
Add helper test method for registering REST calls

We have a pattern in our tests where we tell requests_mock about a URI
we're going to call, then we save a list of the things we want to check
to see have been called. There is more duplicated data thre than I tend
to feel like typing. SO - make a helper method that allows us to
register a call (and optionally some things to check) and save it to the
list of things to assert at the end.

Change-Id: Ie542efd55dc61999c7bad4787427f7664c0ecfb0