Updated openstack/openstack Project: openstack/openstack-manuals  515538be85a815497067776bc0fcc18e07e26620
Fernet tokens need to be cached

This docs suggested that Fernet tokens do not need to be cached at all,
but benchmarking shows that Fernet tokens substantially benefit from
proper caching. The same advice applies to other token types as well, so
I turned this mistake into an intro paragraph for the token caching
section.

Change-Id: I097539ccfeecc3eaa47f61d2b95fa94cb8fc526a