Updated openstack/openstack Project: openstack/python-novaclient  e0d50479ad9b3989639a2fd85544e2243bea3803
x-openstack-request-id logged twice in logs

In the recent release of keystoneauth1 2.18.0 provision is made to log
x-openstack-request-id for session client. Once this new library is synced
in openstack projects, the x-openstack-request-id will be logged twice
on the console if session client is used.

For example,
$ nova --debug list

DEBUG (session:640) GET call to compute for
http://10.232.48.204:8774/v2.1/servers/detail used request id
req-b6aeff56-0408-44ed-80ba-cd7b950a8f21

DEBUG (client:85) GET call to compute for
http://10.232.48.204:8774/v2.1/servers/detail used request id
req-b6aeff56-0408-44ed-80ba-cd7b950a8f21

Above log will be logged twice on the console.

Removed logging of x-openstack-request-id in case of SessionClient as it
is already logged in keystoneauth1.

Depends-On: I492b331ff3da8d0b91178bf0d5fe1d3702f15bd7
Closes-Bug: #1657351

Change-Id: I9ba6a9e498fbcb027f45499d27ece966c52580ba