Updated openstack/openstack Project: openstack-infra/shade  cc78a7fbad1ef7bfb71d0775280d6e891686ab7a
Use port list to find missing floating ips

It's possible for a cloud to have multiple private networks with overlapping
IP ranges. In that case, the check for missing floating ips can
erroneously match a floating ip for a different server.

Ports are actually unique, and are the foreign key between these things.
Instead of starting with list_floating_ips, start with listing the ports
for the server. In the case where OpenStack isn't broken, this will be
the same number of API calls. In the case where it is, there will be one
extra call per server, but ultimately the output will be more correct -
and the fix for the extra load on the cloud is to fix the nova/neutron
port mapping.

Also, fixed the spelling of supplemental.

Story: 2000845
Change-Id: Ie53a2a144ca2ed812d5441868917996f67b6f454