Updated openstack/openstack Project: openstack/cinder  673deb06d1810c8f900a4b1315daf97e34ca51e4
Add lock around SolidFire clone-image

The SolidFire cluster has limits on simultaneous clones
per volume, we've dealt with this pretty easily by just
waiting if we encounter the limit (clones are pretty fast
, so waiting for them isn't a terribly big deal). This
process adds complications to the driver though and can
be racey, so instead we should make this simple and just
use Cinders lock mechanism.

This change adds a lock decorator to the clone_image method
just like we did for the clone_volume method.  This elminates
the need for various tracking and retry mechanisms when using
image caching with SolidFire.

Change-Id: Ia5a034ad578149ddae84ef2fe7e2be95f713072b
Closes-Bug: #1658154