Updated openstack/openstack Project: openstack/nova  35a42c5b12f3b33d0e913f752d31b415ae7f5c86
Fix and add some notes to the cells v2 first time setup doc

There was a typo in here about the map_cell0 command creating
a db connection with a _nova suffix when it's actually a _cell0
suffix like in the example, nova_cell0.

This also adds a reminder to sync the API database schema before
running the commands and also gives a warning about being
specific when using map_cell0 and having the databases on different
hosts.

Depends-On: I541b072638b5d50985145391e76f610417fdcaa6

Change-Id: Ibf3355217bbd0139a020de352bb62ff7d973d27b