Updated openstack/openstack Project: openstack-infra/nodepool  0ebec7cb52ea2cd2097357204ff048116eff2c57
Allow nodepool-builder to only build diskimages

Users can now use nodepool-builder as a front end to
diskimage-builder. The main reason to do this, is we can leverage the
nodepool.yaml file to store the diskimage configuration. Plus
distributed builders.

To actually support this, we need to expose a formats field into the
diskimages section.

Change-Id: I1c0c196828c5c1bf2bf26e278598dbb680d0263e
Signed-off-by: Paul Belanger <pabelanger@redhat.com>