Updated openstack/openstack Project: openstack/heat  b8ae0ae53aff0a7315e1517105a155e5f6bad573
Don't look at convergence_engine conf option in Resource

The only thing that's important to a Resource is whether the *current*
stack is using convergence or not. Whether the default for new stacks is to
use convergence is irrelevant, and using that as the test will result in
unintended behaviour.

Change-Id: I53ca0b5b4bbee173e273cfa625723e61ba2042ef