Updated openstack/openstack Project: openstack/heat  587ffdd6ab2609807265c43b322b8a5919c20144
Always replace DELETE_COMPLETE resources on update

Previously, this only happened in convergence, presumably because it was
assumed that the condition could only show up in convergence. However, even
in the legacy path there exists a brief window between when the Resource is
marked DELETE_COMPLETE and when it is destroyed (removed from the DB). An
exception at that exact point could leave a resource in this state. There's
also people manually fiddling around in the database to contend with.

If a resource is DELETE_COMPLETE then there is no possible way for an
update to do anything sensible with it; we always want to create a
replacement if it still exists in the template.

Change-Id: Ic750f631399084441af95fd1835a136c2c7dc4b4