Updated openstack/openstack Project: openstack-infra/project-config  b63283e583af8ca6528fc39aeca0338c148d3150
Make python3 rally job voting in Neutron

This makes the python3 rally job voting in Neutron and alters
the python2 job to only run on the stable branches.

Note that this job is not in the gate queue even though it is
voting. This is the way the current rally job that it is replacing
is configured.

Change-Id: Ib478da40a5df8f4060f61d251570957fbc363a6a