Updated openstack/openstack Project: openstack/keystone  2fb3fd267c26daf40e0dd8af32e12f0fa2c65724
Remove LDAP delete logic and associated tests

[0][1] made it so that every write operation to the LDAP backend by
users returns a 403 Forbidden. However, since all our unit tests
depend on being able to write to the LDAP backend, the code to
perform writes is still in the codebase and it's toggled by the
unit tests.

This patch cleans up delete operations from unit tests. And since
those unit tests don't depend anymore on being able to do deletes
to LDAP, the associated delete logic in the driver is removed.
None of this is user facing or changes anything in the Keystone
behavior.

[0] I13eada3d5c3a166223c3e3ce70b7054eaed1003a
[1] I225daf0e00742c54d5d009e456d4a3ad864356a0

Change-Id: I9f183492fd318c755ba026e9e402615fe66f100c