Updated openstack/openstack Project: openstack/cinder  964a8d29143b377f3171902fef4e777c66d6fa45
VMware: Allow deletion of snapshot in error state

The VMDK driver fails snapshot deletion if the volume is attached.
There is no need to fail the snapshot deletion if the snapshot does
not exist in the backend (for example, in cases where the snapshot
creation failed and the snapshot is in 'error' state) irrespective
of the volume state. This patch checks the existence of the snapshot
in the backend and returns if there is no backend snapshot.

Closes-bug: #1620697
Change-Id: Id9ad45435e558c4965e3d19801fedcdfa20c50b1