Updated openstack/openstack Project: openstack/python-swiftclient  816565d575489e745e4f8f021791eda2b7774af3
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove python-swiftclient.

Change-Id: I3947a6165eaa9f5cb62a7df4f5a2c16065da2f1d