Updated openstack/openstack Project: openstack-infra/project-config  966b2f3a2bd1d068bfeb532df6590bbf369b07e4
OSA: Make the ceph scenario non voting

Now that the Ceph integration has merged to OSA, we will make the
Ceph scenario test a non-voting check and move it out of
the experimental jobs.

Since this integration was just added in Ocata, the job is
skipped on all branches prior to Ocata.

Change-Id: I684deeae44bed94e223ea4b5ac6f9f2b30d89262