Updated openstack/openstack Project: openstack/nova  66c661258873e2544e286099c4bc027c26c851c4
tests: Replace use of CONF with monkey patching

Per the docs, the 'nova-console' service is "a XenAPI-specific service
that most recent VNC proxy architectures do not use." [1] This service
provides a config option - 'console_driver' - which only has two valid
values, one of which was a fake used in testing.

There are better ways to stub out code during tests that by setting
configuration options, namely through the use of monkey patching. Do
this, paving the way for deprecation and eventual removal of this
configuration option.

[1] https://github.com/openstack/openstack-manuals/tree/4b1e8f1/doc\
    /admin-guide/source/compute-remote-console-access.rst

Change-Id: Ibfc957c40dafb1ecb35b8658ab92b922b1f39a1a