Updated openstack/openstack Project: openstack/ironic-python-agent  44832822fe0262c8fae8a041c8c10ed4f1f320f3
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Ic86a4c837704cdde2f9d0051ea2f169015816ab7