Updated openstack/openstack Project: openstack/cinder  275be41863e189223fa64bc9eabb45fbb171cd74
Gate migration tests: Add Cinder tempest hook

This patch is part of the effort to add migration tests to OpenStack
gate. It will rely in 2 tempest tests that are on the dependency list and that
the gate job will configure the following backends: ceph nfs glusterfs lvm.
Once running, the script will configure 2 tempest backends and run the
migration tests on all possible combinations of the backend listed above. The
output of the test run will be something like:

                    ceph             nfs
    ceph             ---         SUCCESS
     nfs         SUCCESS             ---

If all tests in the matrix result in SUCCESS, then the test script returns.
Otherwise, other value is returned and the job is marked as failed.

Depends-on: I4ed9b1f30d4e4e595c44a0ce243c2463069833d7
Depends-on: Iff4d880a0f8e928cbc07b84ff0c357ad59ef929c
Change-Id: I17b5cd9dafb3ad17e8721376ae1e6790bd93108c