Updated openstack/openstack Project: openstack/glance  fdf02374994efe76a89651d296f97e557e0ccea3
Replace master/slave usage in replication

The replicator replicates from a source to a target. There is no
other relationship or responsibilities between the source and target
but to provide data and to receive data. The target is not a slave.
Source and Target fits each more accurately.

This changes the configuration parameter without using the standard
deprecation process because the short options cannot be deprecated
using that method. Related logging messages are adjusted to match.

No new short options are added at this time due to conflicts in use
of S. At a later time if deprecated names are removed the short names
for source and target can be introduced without conflict.

At a later time if deprecated names are removed the short names for
source and target can be introduced without conflict.

Change-Id: Iab9dcc0e3de30a3a98a659a6bb32ac3604d27001