Updated openstack/openstack Project: openstack/neutron  1f5f10e4c8ebe8e9c0861fd3cd67522fd0109ac1
Manually add pk for alembic_version table

We hit a problem that Neutron migrations cannot be applied
on galera cluster with ENFORCING mode as alembic_version table
missing primary key on the table and it is expected that all tables
will have primary keys.

alembic_version is created and managed by Alembic and fixes for this
on Alembic side are on the review [1]. Anyway here migration is
proposed that will add pk on this table if it is missing, as in
other case we will have inconsistency between new deployments and
old ones.

[1] - https://gerrit.sqlalchemy.org/#/c/279/

Change-Id: I543a1ee286bdf11ae35adb87125d044a351a2648
Closes-bug: #1655610